+++ wiki/Changelog.md
@@ -14,6 +14,7 @@ Rendering
   * Speedup of width detection for auto-narrowing for certain characters (#1065, #979, #892).
   * Prevent artefacts of large-size underscore cursor (CSI 4 SP q CSI ? 6 c).
   * Prevent spacing anomaly after U+FD3E and U+FD3F.
+  * Fix emojis selection highlighting (#1116), reverting 3.0.1 tweak for emojis in bidi lines.
 
 Window handling
   * Ensure -w full to cover taskbar also with -B void (~#1114).